# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistem Beasiswa API Dokumentasi",
        "description": "Dokumentasi API untuk Sistem Beasiswa Universitas Esa Unggul",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/dokumen": {
            "get": {
                "tags": [
                    "Dokumen"
                ],
                "summary": "Menampilkan semua dokumen",
                "operationId": "d46181f21fafa4088995a53e6e2ee0f6",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar dokumen"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dokumen"
                ],
                "summary": "Upload dokumen baru",
                "operationId": "6e6c0838e6ba7aed119dbc6ae5068c9e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "pendaftaran_id"
                                ],
                                "properties": {
                                    "pendaftaran_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "ktm": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "krs": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "khs": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dokumen berhasil disimpan"
                    }
                }
            }
        },
        "/api/dokumen/{id}": {
            "get": {
                "tags": [
                    "Dokumen"
                ],
                "summary": "Menampilkan detail dokumen",
                "operationId": "908a8ac65361a3712f180c2975189c33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail dokumen"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dokumen"
                ],
                "summary": "Memperbarui dokumen (upload ulang)",
                "operationId": "d2cc804a4be97caceb124e0f02320936",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "ktm": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "krs": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "khs": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dokumen berhasil diperbarui"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dokumen"
                ],
                "summary": "Menghapus dokumen",
                "operationId": "ef516848303f13421dc5e3bfed775da2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Dokumen berhasil dihapus"
                    }
                }
            }
        },
        "/api/pendaftaran": {
            "get": {
                "tags": [
                    "Pendaftaran"
                ],
                "summary": "Menampilkan semua data pendaftaran beasiswa",
                "operationId": "79b44847ba89b758b7d4ccf1826628bd",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data pendaftaran"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pendaftaran"
                ],
                "summary": "Mendaftarkan pengguna ke program beasiswa",
                "operationId": "c9c1ac52f5579de30fc305f7a0908421",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "beasiswa_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "beasiswa_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "status_verifikasi": {
                                        "type": "string",
                                        "example": "diproses"
                                    },
                                    "catatan_verifikasi": {
                                        "type": "string",
                                        "example": "Dokumen belum lengkap"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pendaftaran berhasil disimpan"
                    }
                }
            }
        },
        "/api/pendaftaran/{id}": {
            "get": {
                "tags": [
                    "Pendaftaran"
                ],
                "summary": "Menampilkan detail pendaftaran berdasarkan ID",
                "operationId": "9c8d0d5f4bff4f709151654b74f13bce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail pendaftaran ditemukan"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pendaftaran"
                ],
                "summary": "Memperbarui status dan catatan verifikasi pendaftaran",
                "operationId": "6fec343d9f45ded326294e194a8aa3ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status_verifikasi": {
                                        "type": "string",
                                        "example": "disetujui"
                                    },
                                    "catatan_verifikasi": {
                                        "type": "string",
                                        "example": "Dokumen lengkap"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Data pendaftaran berhasil diperbarui"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pendaftaran"
                ],
                "summary": "Menghapus data pendaftaran",
                "operationId": "e5aa2fb7d2f3d1a4cadcdfc1d65f981c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pendaftaran berhasil dihapus"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}